Run auction on each update call
Problem: We have an emulated global state of the auction in is20
token that should be updated as soon as each update call queries.
For instance we want to required to top up cycles balance of the
auction as soon as possible. This now can be done via the
`pre_update` hook. We now can update auction state each time we
query any update call by injecting the code via `update` macro.
Note, that the error will be ignored when we inject the code
because we don't want any update call fail even if auction fails.