Add API method `state_check`
This mehods allows getting the current state type and version from the
canister (both token and factory). It is intended to be used for CI
pre-upgrade checks to verify that the state was not modified in an
unexpected way.

I also removed the main.rs from `is20-token` crate for this create is
not supposed to be built as a canister any more.